don't use a border-box for the display-inner div
commit 5a4bd42557a "Fix VNC canvas stretch" fixed resizing the canvas
by compensating for the border width. This looks a bit odd, since the
div is first set to be a border-box, which has the border inside the
div and thereafter the size is corrected to undo that again. So simply
don't make it a border-box and use the correct size.