add optional https support
Adding proper TSL support is rather difficult since the majority of the installations
are not connected to the internet at all, and hence it is simply impossible to renew a
signed certificate. Unfortunately browsers warn so much nowadays about self signed
certificates that it is simply impossible to deploy that in a large scale, since it will
raise loads of issue if the browser tells you it is unsafe, not secure etc etc. (while
it is actually more secure then plain text over http).

So this adds support to encrypt the websocket connection, but only if you managed to
connect with https in the first place.

https://github.com/victronenergy/venus/issues/464